Temporarily disabling CI of zenoh transport
This breaks a dependency loop caused by the bugfix for up-cpp being
merged to main. The zenoh transport has a PR that takes in those
changes, but it is blocked waiting on 1.0.1-rc1 of up-cpp. The PR adding
up-cpp 1.0.1-rc1 is blocked because main in up-transport-zenoh-cpp
cannot build as-is against up-cpp main.