Relabeling current up-core-api version
This change is to better align with the upstream tags in up-spec. When
we selected the version numbering here, we had expected semantic
versioning to occur on the uProtocol spec repo. However, instead only
the alpha number is incrementing with each release. As such, we need to
change our version numbers to match.

This change will require developers to `conan remove 'up-core-api'`
before rebuilding all packages that depend on up-core-api.